Check that an overlay is visible before removing
Add a guard to ensure that before `_hideOverlay` removes an overlay that
there is currently a visible entry of that overlay type. This resolves
an issue which would cause the flutter engine to throw as
`AssertionError` if either `hideLoader` or `hideNotification` are called
whilst no overlay of that type is currently showing.

closes #16